+++ changelog.md
@@ -15,6 +15,4 @@
 * New logo
 
 # v1.0.0
-
- * Initial release
- With such features as: The entire thing working perfectly with 2.2 levels but not on 2.1
+* Initial release with such features as: The entire thing working perfectly with 2.2 levels but not on 2.1