+++ changelog.md
@@ -1,5 +1,6 @@
 # v1.2.0
 * Added setting for maximum Xs
+* Added cat
 
 # v1.1.2
 * Fixed occasional-er crashes