+++ changelog.md
@@ -1,6 +1,6 @@
 # v1.2.0
 * Added setting for maximum Xs
-* Added cat
+* Added power notation (disabled by default)
 
 # v1.1.2
 * Fixed occasional-er crashes